Adds some supervision to all this
All UserTracker processes are registered with the Analytics.Registry
so that we can access a global list of all of the UserTracker processes.

These processes are also managed by a DynamicSupervisor so that they
are not linked to the channel process that started them.

The idea is that these processes should live on when the user navigates
away from their current page, and even arguably after they leave the
site.  It should stay alive for long enough to still be there if they
happen to come back within a reasonable amount of time (about 1 hour).