Renames the channel to something more descriptive
It's also less nefarious this way, since users will be able to see that
they're joining a channel for analytics purposes.  Considering this is
all server-side analytics that aren't sent to a third party, it's
pretty innocuous, comparatively.

This also adds pageviews to the existing UserTracker, if it already exists