Lowers tracking threshold to ten minutes
This also improves the session id cookie to reset its expiration on every
page load.  This way, the cookie and tracking process should be roughly
aligned.

The idea here is that if the user is inactive for ten minutes, then their
next activity should be treated as a new session.